SLING-2944 Revert Sling API dependency to 2.2.0 again
Sling API 2.4.0 is not required and probably only has been
updated to make sure the import version range for the Resource
API is correct. Given SLING-2993 we should actually provide
proper annotation of implemented API for the bundle plugin
to properly devise the import version range.

For now removing the 'provide:=true' import tag solves this
issue, since we only implement the ResourceProvider interface
(intended to be @ConsumerType) and extend AbstractSlingResource
(which is safe to extend in @ConsumerType fashion).

git-svn-id: https://svn.apache.org/repos/asf/sling/trunk@1510567 13f79535-47bb-0310-9956-ffa450edef68